refactor(components): move log out feature to new button
